refactor code style
