+++ easy_rec/python/model/ppnet_v3m.py
@@ -6,10 +6,7 @@ import logging
 import tensorflow as tf
 from tensorflow.python.framework import ops
 
-from easy_rec.python.layers import dnn
-from easy_rec.python.layers import mmoe
 from easy_rec.python.model.keep_model.model_ps_mmoe import CustomizedModel
-from easy_rec.python.model.multi_task_model import MultiTaskModel
 from easy_rec.python.model.rank_model import RankModel
 from easy_rec.python.protos.ppnet_pb2 import PPNet as PPNetConfig
 
@@ -71,21 +68,20 @@ class PPNetV3M(RankModel):
       #      tf.reduce_mean(output), tf.reduce_min(self._labels[lbl_name]),
       #      tf.reduce_max(self._labels[lbl_name]),
       #      tf.reduce_mean(self._labels[lbl_name])], message='output')
-      metric_dict['auc_' + lbl_name]  = metrics_tf.auc(self._labels[lbl_name],
-          output, num_thresholds=1000) 
+      metric_dict['auc_' + lbl_name] = metrics_tf.auc(
+          self._labels[lbl_name], output, num_thresholds=1000)
     return metric_dict
-     
 
   def build_loss_graph(self):
     for lbl_id in range(len(self._model_conf['label'])):
       lbl_info = self._model_conf['label'][lbl_id]
       lbl_name = lbl_info.get('input_name')
       output = self._prediction_dict.get(lbl_name)
-      tf.summary.scalar('label/%s' % lbl_name, 
-               tf.reduce_mean(tf.to_float(self._labels[lbl_name])))
+      tf.summary.scalar('label/%s' % lbl_name,
+                        tf.reduce_mean(tf.to_float(self._labels[lbl_name])))
       tf.summary.scalar('predict/%s' % lbl_name, tf.reduce_mean(output))
-      loss_obj = tf.keras.losses.BinaryCrossentropy(reduction='sum_over_batch_size')(
-          self._labels[lbl_name], output)
+      loss_obj = tf.keras.losses.BinaryCrossentropy(
+          reduction='sum_over_batch_size')(self._labels[lbl_name], output)
       # loss_obj = tf.Print(loss_obj, [tf.reduce_min(loss_obj), tf.reduce_max(loss_obj),
       #        tf.reduce_mean(loss_obj), tf.shape(loss_obj)], message='loss_obj')
       self._loss_dict[lbl_name] = loss_obj
